refactor(tui): split model lifecycle helpers into focused modules
Keep the public reducer and coordinated turn finalization unchanged while extracting block creation, inspection grouping, navigation, shell tracking, and tool-result presentation.
