+++ packages/agent-tui/src/Agent/TUI/Model.hs
@@ -47,7 +47,6 @@ import Agent.TUI.Presentation
     ( formatToolDiffRelative
     , formatToolDiffRelativeWithOutput
     , formatToolOutputRelative
-    , isInspectionTool
     , todoListFromToolArguments
     , todoListFromToolOutput
     , toolCallHeaderRelative
@@ -55,13 +54,17 @@ import Agent.TUI.Presentation
     , toolCallTitleRelative
     )
 import Agent.TUI.TextWidth (clampGraphemeCursor)
+import Agent.TUI.Model.Block
 import Agent.TUI.Model.Edit
+import Agent.TUI.Model.Inspection
+import Agent.TUI.Model.Selection
+import Agent.TUI.Model.Shell
+import Agent.TUI.Model.ToolResult
 import Agent.TUI.Model.State
 import Agent.TUI.Model.Timing
 import Agent.TUI.Model.Types
 import Agent.TUI.Motion
     ( completionStatusDurationMillis )
-import qualified Agent.Json.Decode as Hermes
 import Agent.Loop
     ( LoopEvent(..)
     , TokenUsage(..)
@@ -70,9 +73,7 @@ import Agent.Loop
     )
 import Agent.Telemetry (telemetrySummary)
 import Agent.ToolDispatch
-    ( ToolOutcome(..)
-    , toolCallResultOutcome
-    , ToolCall(..)
+    ( ToolCall(..)
     , ToolCallResult(..)
     , canonicalToolName
     )
@@ -85,7 +86,6 @@ import Data.Sequence (Seq)
 import qualified Data.Sequence as Seq
 import Data.Text (Text)
 import qualified Data.Text as Text
-import Data.Char (isSpace)
 
 reduceUi :: UiEvent -> UiState -> UiState
 reduceUi event state = case event of
@@ -596,259 +596,6 @@ startToolCall call state
                         state.uiToolCalls
                 }
 
-startInspectionCall :: ToolCall -> UiState -> UiState
-startInspectionCall call state =
-    case Seq.viewr state.uiBlocks of
-        _ Seq.:> block
-            | block.blockKind == BlockInspect
-            , Just group <- Map.lookup block.blockId state.uiInspectionGroups
-            , group.inspectionGroupOpen ->
-                extend block group
-        _ -> start
-  where
-    activity = toolCallTitleRelative state.uiWorkspaceRoot call
-    (title, headerDetail) =
-        toolCallHeaderRelative state.uiWorkspaceRoot call
-    detail = fromMaybe "" headerDetail
-    item = InspectionItem
-        { inspectionCallId = call.callId
-        , inspectionToolName = canonicalToolName call.name
-        , inspectionTitle = title
-        , inspectionDetail = detail
-        , inspectionBody =
-            formatToolDiffRelative state.uiWorkspaceRoot call
-        , inspectionState = BlockRunning
-        }
-    common current =
-        current
-            { uiRunning = True
-            , uiGenerating = False
-            , uiAwaitingInput = False
-            , uiActivity = activity
-            }
-    extend block group =
-        let
-            blockIndex = Seq.length state.uiBlocks - 1
-            updatedGroup =
-                group
-                    { inspectionGroupItems =
-                        group.inspectionGroupItems <> [item]
-                    }
-        in common state
-            { uiBlocks =
-                Seq.adjust
-                    (renderInspectionGroup updatedGroup)
-                    blockIndex
-                    state.uiBlocks
-            , uiInspectionGroups =
-                Map.insert
-                    block.blockId
-                    updatedGroup
-                    state.uiInspectionGroups
-            , uiToolCalls =
-                Map.insert
-                    call.callId
-                    (blockIndex, call)
-                    state.uiToolCalls
-            }
-    start =
-        let
-            prepared = closeInspectionGroups state
-            blockIndex = Seq.length prepared.uiBlocks
-            ident = BlockId prepared.uiNextBlockId
-            group = InspectionGroup
-                { inspectionGroupOpen = True
-                , inspectionGroupItems = [item]
-                }
-            appended =
-                appendBlock
-                    BlockInspect
-                    title
-                    item.inspectionBody
-                    detail
-                    BlockRunning
-                    (Just call.callId)
-                    (common prepared)
-        in appended
-            { uiBlocks =
-                Seq.adjust
-                    (\block ->
-                        block { blockInspectionGroupable = True })
-                    blockIndex
-                    appended.uiBlocks
-            , uiInspectionGroups =
-                Map.insert ident group appended.uiInspectionGroups
-            , uiToolCalls =
-                Map.insert
-                    call.callId
-                    (blockIndex, call)
-                    appended.uiToolCalls
-            }
-
--- | Close a burst when another visible event intervenes. Keep its item
--- metadata until the turn ends so a provider can still retract an individual
--- completed call without removing the rest of the grouped block.
-closeInspectionGroups :: UiState -> UiState
-closeInspectionGroups state =
-    state
-        { uiInspectionGroups =
-            Map.map
-                (\group -> group { inspectionGroupOpen = False })
-                state.uiInspectionGroups
-        }
-
-renderInspectionGroup :: InspectionGroup -> UiBlock -> UiBlock
-renderInspectionGroup group block =
-    block
-        { blockTitle = inspectionGroupTitle items
-        , blockBody = inspectionGroupBody items
-        , blockState = inspectionGroupState items
-        , blockDetail = inspectionGroupDetail items
-        , blockCallId = (.inspectionCallId) <$> listToMaybe items
-        , blockInspectionGroupable =
-            block.blockInspectionGroupable && length items == 1
-        }
-  where
-    items = group.inspectionGroupItems
-
-inspectionGroupTitle :: [InspectionItem] -> Text
-inspectionGroupTitle [] = "Inspected"
-inspectionGroupTitle [item] = item.inspectionTitle
-inspectionGroupTitle items =
-    Text.intercalate ", " (inspectionSummaries items)
-        <> statusSuffix
-  where
-    failed =
-        length
-            (filter
-                ((== BlockFailed) . (.inspectionState))
-                items)
-    denied =
-        length
-            (filter
-                ((== BlockDenied) . (.inspectionState))
-                items)
-    suffixes =
-        [ Text.pack (show failed) <> " failed" | failed > 0 ]
-            <> [ Text.pack (show denied) <> " denied" | denied > 0 ]
-    statusSuffix =
-        case suffixes of
-            [] -> ""
-            _ -> " · " <> Text.intercalate ", " suffixes
-
-inspectionGroupDetail :: [InspectionItem] -> Text
-inspectionGroupDetail [item] = item.inspectionDetail
-inspectionGroupDetail _ = ""
-
-inspectionSummaries :: [InspectionItem] -> [Text]
-inspectionSummaries =
-    map render . foldl add []
-  where
-    add counts item =
-        let label = inspectionSummaryLabel item.inspectionToolName
-        in case break ((== label) . fst) counts of
-            (before, (_, count) : after) ->
-                before <> [(label, count + 1)] <> after
-            _ -> counts <> [(label, 1)]
-    render :: (Text, Int) -> Text
-    render (label, count) =
-        label
-            <> " "
-            <> Text.pack (show count)
-            <> if count == 1 then " item" else " items"
-
-inspectionSummaryLabel :: Text -> Text
-inspectionSummaryLabel name
-    | name `elem` ["read_file", "read_tool_output", "mcp_read_resource"] =
-        "Read"
-    | name
-        `elem` ["list_dir", "mcp_list_resources", "list_agents", "ListAgents", "Glob"] =
-        "Listed"
-    | name
-        `elem`
-            [ "grep"
-            , "search_tool_output"
-            , "mcp_search"
-            , "search_tool"
-            , "conversation_search"
-            , "skill_search"
-            , "WebSearch"
-            , "ToolSearch"
-            ] =
-        "Searched"
-    | otherwise = "Inspected"
-
--- Image rendering and long-running task-output polling attach lifecycle data
--- to one exact block, so only compact read/list/search calls join a burst.
-isGroupableInspectionTool :: Text -> Bool
-isGroupableInspectionTool rawName =
-    canonicalToolName rawName
-        `elem`
-            [ "read_file"
-            , "list_dir"
-            , "grep"
-            , "read_tool_output"
-            , "search_tool_output"
-            , "mcp_search"
-            , "search_tool"
-            , "mcp_list_resources"
-            , "mcp_read_resource"
-            , "conversation_search"
-            , "skill_search"
-            , "view_skill"
-            , "list_agents"
-            , "Glob"
-            , "WebSearch"
-            , "ToolSearch"
-            , "ListAgents"
-            ]
-
-inspectionGroupBody :: [InspectionItem] -> Text
-inspectionGroupBody [item] = item.inspectionBody
-inspectionGroupBody items =
-    Text.intercalate "\n" headers
-        <> if null details
-            then ""
-            else "\n\n" <> Text.intercalate "\n\n" details
-  where
-    headers = map inspectionItemHeader items
-    details =
-        [ inspectionItemTitle item <> "\n"
-            <> Text.unlines
-                (map ("    " <>) (Text.lines item.inspectionBody))
-        | item <- items
-        , not (Text.null (Text.strip item.inspectionBody))
-        ]
-
-inspectionItemHeader :: InspectionItem -> Text
-inspectionItemHeader item =
-    "  "
-        <> inspectionStateGlyph item.inspectionState
-        <> " "
-        <> inspectionItemTitle item
-
-inspectionItemTitle :: InspectionItem -> Text
-inspectionItemTitle item
-    | Text.null item.inspectionDetail = item.inspectionTitle
-    | otherwise = item.inspectionTitle <> " " <> item.inspectionDetail
-
-inspectionStateGlyph :: BlockState -> Text
-inspectionStateGlyph = \case
-    BlockComplete -> "◇"
-    BlockFailed -> "✗"
-    BlockCancelled -> "⊘"
-    BlockDenied -> "⊘"
-    BlockStreaming -> "◆"
-    BlockRunning -> "◆"
-
-inspectionGroupState :: [InspectionItem] -> BlockState
-inspectionGroupState items
-    | any ((== BlockRunning) . (.inspectionState)) items = BlockRunning
-    | any ((== BlockFailed) . (.inspectionState)) items = BlockFailed
-    | any ((== BlockDenied) . (.inspectionState)) items = BlockDenied
-    | any ((== BlockCancelled) . (.inspectionState)) items = BlockCancelled
-    | otherwise = BlockComplete
-
 replaceOrAppendRecap :: Text -> BlockState -> UiState -> UiState
 replaceOrAppendRecap body blockState state =
     case latestRecapIndex state of
@@ -872,10 +619,6 @@ latestRecapIndex state =
     case Seq.findIndexR ((== BlockRecap) . (.blockKind)) state.uiBlocks of
         Just index -> Just index
         Nothing -> Nothing
-
-selectedIndexFor :: Maybe BlockId -> Seq UiBlock -> Maybe Int
-selectedIndexFor selected remaining =
-    selected >>= \ident -> Seq.findIndexL ((== ident) . (.blockId)) remaining
 
 removeBlockAt :: Int -> UiState -> UiState
 removeBlockAt index state =
@@ -922,42 +665,6 @@ removeBlockAt index state =
         , uiShellPolls = retainShellPolls processes state.uiShellPolls
         , uiTurnStartBlock = adjustIndex state.uiTurnStartBlock
         , uiAttemptStartBlock = adjustIndex state.uiAttemptStartBlock
-        }
-
-appendBlock
-    :: BlockKind
-    -> Text
-    -> Text
-    -> Text
-    -> BlockState
-    -> Maybe Text
-    -> UiState
-    -> UiState
-appendBlock kind title body detail blockState callId state =
-    let prepared = closeInspectionGroups state
-        index = Seq.length prepared.uiBlocks
-        ident = BlockId prepared.uiNextBlockId
-        block = UiBlock
-            { blockId = ident
-            , blockKind = kind
-            , blockTitle = title
-            , blockBody = body
-            , blockTimestamp = ""
-            , blockDetail = detail
-            , blockState
-            , blockExpanded =
-                kind `elem` [BlockUser, BlockAssistant, BlockSystem, BlockRecap, BlockError]
-                    || (kind == BlockShell
-                        && blockState `elem` [BlockStreaming, BlockRunning])
-            , blockCallId = callId
-            , blockInspectionGroupable = False
-            }
-    in prepared
-        { uiBlocks = prepared.uiBlocks Seq.|> block
-        , uiNextBlockId = prepared.uiNextBlockId + 1
-        , uiSelectedBlock = Just ident
-        , uiSelectedBlockIndex = Just index
-        , uiBlockIndices = Map.insert ident index prepared.uiBlockIndices
         }
 
 -- | Attach one captured wall-clock label to newly appended conversation
@@ -1096,139 +803,6 @@ finishVisibleTool result state =
             | blockIndex < Seq.length state.uiBlocks ->
                 completeTool blockIndex call displayed next
         Just _ -> next
-
-trackedShellOwner :: ToolCall -> UiState -> Maybe BlockId
-trackedShellOwner call state = do
-    sessionId <- writeStdinSession call
-    Map.lookup sessionId state.uiShellProcesses
-
--- A non-empty write_stdin remains visible as an input action, but it can also
--- observe the underlying command finishing. Keep the process owner in sync
--- without duplicating the continuation output across both blocks.
-reconcileVisibleShellContinuation
-    :: ToolCall
-    -> ToolCallResult
-    -> UiState
-    -> UiState
-reconcileVisibleShellContinuation call result state =
-    case writeStdinSession call of
-        Nothing -> state
-        Just sessionId ->
-            case Map.lookup sessionId state.uiShellProcesses of
-                Nothing -> state
-                Just ownerId ->
-                    case runningShellOutcome result of
-                        Just (returnedSessionId, _) ->
-                            state
-                                { uiShellProcesses =
-                                    Map.insert
-                                        returnedSessionId
-                                        ownerId
-                                        (Map.delete
-                                            sessionId
-                                            state.uiShellProcesses)
-                                }
-                        Nothing ->
-                            case terminalShellOutcome result of
-                                Nothing -> state
-                                Just (blockState, _) ->
-                                    let
-                                        processes =
-                                            Map.delete
-                                                sessionId
-                                                state.uiShellProcesses
-                                        blocks =
-                                            case lookupBlockIndex ownerId state of
-                                                Nothing -> state.uiBlocks
-                                                Just (blockIndex, _) ->
-                                                    Seq.adjust
-                                                        (setBlockState blockState)
-                                                        blockIndex
-                                                        state.uiBlocks
-                                    in state
-                                        { uiBlocks = blocks
-                                        , uiShellProcesses = processes
-                                        , uiShellPolls =
-                                            retainShellPolls
-                                                processes
-                                                state.uiShellPolls
-                                        }
-
-retainRunningShell :: Int -> Int -> Text -> UiState -> UiState
-retainRunningShell blockIndex sessionId output state =
-    case Seq.lookup blockIndex state.uiBlocks of
-        Nothing -> state
-        Just owner ->
-            state
-                { uiBlocks =
-                    Seq.adjust
-                        (\block ->
-                            setBlockState BlockRunning block
-                                { blockBody = output })
-                        blockIndex
-                        state.uiBlocks
-                , uiShellProcesses =
-                    Map.insert
-                        sessionId
-                        owner.blockId
-                        state.uiShellProcesses
-                }
-
-finishShellPoll :: Int -> ToolCallResult -> UiState -> UiState
-finishShellPoll sessionId result state =
-    let
-        next =
-            state
-                { uiRunning = True
-                , uiAwaitingInput = False
-                , uiActivity = "Thinking…"
-                , uiShellPolls =
-                    Map.delete result.callId state.uiShellPolls
-                }
-        withoutSession =
-            Map.delete sessionId next.uiShellProcesses
-    in case Map.lookup sessionId state.uiShellProcesses of
-        Nothing ->
-            next { uiShellProcesses = withoutSession }
-        Just ownerId ->
-            case lookupBlockIndex ownerId next of
-                Nothing ->
-                    next { uiShellProcesses = withoutSession }
-                Just (blockIndex, _) ->
-                    case runningShellOutcome result of
-                        Just (returnedSessionId, output) ->
-                            appendShellOutput blockIndex output BlockRunning
-                                next
-                                    { uiShellProcesses =
-                                        Map.insert
-                                            returnedSessionId
-                                            ownerId
-                                            withoutSession
-                                    }
-                        Nothing ->
-                            case terminalShellOutcome result of
-                                Nothing -> next
-                                Just (blockState, output) ->
-                                    appendShellOutput
-                                        blockIndex
-                                        output
-                                        blockState
-                                        next
-                                            { uiShellProcesses =
-                                                withoutSession
-                                            }
-
-appendShellOutput :: Int -> Text -> BlockState -> UiState -> UiState
-appendShellOutput blockIndex output blockState state =
-    state
-        { uiBlocks =
-            Seq.adjust
-                (\block ->
-                    setBlockState blockState block
-                        { blockBody = block.blockBody <> output })
-                blockIndex
-                state.uiBlocks
-        }
 
 finalizeAttempt :: BlockState -> UiState -> UiState
 finalizeAttempt terminalState state =
@@ -1617,21 +1191,6 @@ finalizeTurn terminalState state =
         , uiShellPolls = Map.empty
         }
 
--- Shell commands start open while output is live, then compact to a one-line
--- summary. Preserve any manual folding while they are still running; a user
--- can also expand the completed block again with the normal block toggle.
-setBlockState :: BlockState -> UiBlock -> UiBlock
-setBlockState blockState block =
-    block
-        { blockState
-        , blockExpanded =
-            if block.blockKind == BlockShell
-                && block.blockState `elem` [BlockStreaming, BlockRunning]
-                && blockState `notElem` [BlockStreaming, BlockRunning]
-                then False
-                else block.blockExpanded
-        }
-
 infoNotice, successNotice, warningNotice, progressNotice, errorNotice
     :: Text -> UiNotice
 infoNotice = transientNotice NoticeInfo
@@ -1671,271 +1230,3 @@ hasAssistantTextSince start =
                 && not (Text.null (Text.strip block.blockBody)))
         . Seq.drop start
         . (.uiBlocks)
-
-moveSelection :: Int -> UiState -> UiState
-moveSelection delta state =
-    case Seq.lookup next blocks of
-        Nothing ->
-            state
-                { uiSelectedBlock = Nothing
-                , uiSelectedBlockIndex = Nothing
-                }
-        Just block ->
-            state
-                { uiSelectedBlock = Just block.blockId
-                , uiSelectedBlockIndex = Just next
-                , uiFollow = next == lastIndex
-                }
-  where
-    blocks = state.uiBlocks
-    lastIndex = Seq.length blocks - 1
-    next = max 0 (min lastIndex (selectedBlockIndex state + delta))
-
-selectBlock :: BlockId -> UiState -> UiState
-selectBlock ident state =
-    case lookupBlockIndex ident state of
-        Nothing -> state
-        Just (index, _) ->
-            state
-                { uiSelectedBlock = Just ident
-                , uiSelectedBlockIndex = Just index
-                , uiFollow =
-                    index == Seq.length state.uiBlocks - 1
-                }
-
-lookupBlock :: BlockId -> UiState -> Maybe UiBlock
-lookupBlock ident state =
-    snd <$> lookupBlockIndex ident state
-
-lookupBlockIndex :: BlockId -> UiState -> Maybe (Int, UiBlock)
-lookupBlockIndex ident state = do
-    index <- Map.lookup ident state.uiBlockIndices
-    block <- Seq.lookup index state.uiBlocks
-    if block.blockId == ident
-        then Just (index, block)
-        else Nothing
-
-selectedBlockIndex :: UiState -> Int
-selectedBlockIndex state =
-    maybe fallback fst (selectedBlockEntry state)
-  where
-    fallback = max 0 (Seq.length state.uiBlocks - 1)
-
-toggleSelected :: UiState -> UiState
-toggleSelected state =
-    case selectedBlockEntry state of
-        Nothing -> state
-        Just (index, _) ->
-            state
-                { uiBlocks =
-                    Seq.adjust
-                        (\block ->
-                            block
-                                { blockExpanded = not block.blockExpanded })
-                        index
-                        state.uiBlocks
-                }
-
-selectedBlockEntry :: UiState -> Maybe (Int, UiBlock)
-selectedBlockEntry state = do
-    ident <- state.uiSelectedBlock
-    index <- state.uiSelectedBlockIndex
-    storedIndex <- Map.lookup ident state.uiBlockIndices
-    block <- Seq.lookup index state.uiBlocks
-    if storedIndex == index && block.blockId == ident
-        then Just (index, block)
-        else Nothing
-
-selectionAfterTruncate
-    :: Seq UiBlock
-    -> Maybe Int
-    -> Maybe (Int, UiBlock)
-selectionAfterTruncate blocks selected =
-    case selected of
-        Just index
-            | Just block <- Seq.lookup index blocks ->
-                Just (index, block)
-        _ ->
-            let index = Seq.length blocks - 1
-            in (\block -> (index, block)) <$> Seq.lookup index blocks
-
-retainShellProcesses
-    :: Seq UiBlock
-    -> Map.Map Int BlockId
-    -> Map.Map Int BlockId
-retainShellProcesses blocks =
-    Map.filter \ownerId ->
-        any
-            (\block ->
-                block.blockId == ownerId
-                    && block.blockState == BlockRunning)
-            blocks
-
-retainShellPolls
-    :: Map.Map Int BlockId
-    -> Map.Map Text Int
-    -> Map.Map Text Int
-retainShellPolls processes =
-    Map.filter (`Map.member` processes)
-
-writeStdinInput :: ToolCall -> Maybe (Int, Maybe Text)
-writeStdinInput call = do
-    guard (canonicalToolName call.name == "write_stdin")
-    either (const Nothing) Just $
-        Hermes.decodeText
-            (Hermes.object $
-                (,)
-                    <$> Hermes.atKey "session_id" Hermes.int
-                    <*> Hermes.optionalKey "chars" Hermes.text)
-            call.arguments
-
-writeStdinSession :: ToolCall -> Maybe Int
-writeStdinSession = fmap fst . writeStdinInput
-
-emptyWriteStdinSession :: ToolCall -> Maybe Int
-emptyWriteStdinSession call = do
-    (sessionId, chars) <- writeStdinInput call
-    guard (maybe True Text.null chars)
-    pure sessionId
-
--- Older/imported results have no execution facts and use the legacy adapter.
-resultBlockState :: ToolCallResult -> BlockState
-resultBlockState result = maybe (toolResultState result.output) outcomeBlockState
-    (toolCallResultOutcome result)
-
-outcomeBlockState :: ToolOutcome -> BlockState
-outcomeBlockState = \case
-    ToolSucceeded -> BlockComplete
-    ToolFailed -> BlockFailed
-    ToolDenied -> BlockDenied
-    ToolCancelled -> BlockCancelled
-    ShellRunning _ -> BlockRunning
-    ShellExited 0 -> BlockComplete
-    ShellExited _ -> BlockFailed
-    ShellCancelled -> BlockCancelled
-    ShellTimedOut -> BlockFailed
-
-runningShellOutcome :: ToolCallResult -> Maybe (Int, Text)
-runningShellOutcome result = case toolCallResultOutcome result of
-    Just (ShellRunning sessionId) -> Just (sessionId, dropOutputLine (dropOutputLine result.output))
-    Just _ -> Nothing
-    Nothing -> runningShellResult result.output
-
-terminalShellOutcome :: ToolCallResult -> Maybe (BlockState, Text)
-terminalShellOutcome result = case toolCallResultOutcome result of
-    Just outcome@(ShellExited _) -> Just (outcomeBlockState outcome, dropOutputLine result.output)
-    Just ShellCancelled -> Just (BlockCancelled, result.output)
-    Just ShellTimedOut -> Just (BlockFailed, result.output)
-    Just _ -> Nothing
-    Nothing -> terminalShellResult result.output
-
--- Only remove the transport prologue for presentation. It is never parsed for
--- execution status, exit codes or session identifiers.
-dropOutputLine :: Text -> Text
-dropOutputLine = Text.drop 1 . Text.dropWhile (/= '\n')
-
-runningShellResult :: Text -> Maybe (Int, Text)
-runningShellResult output = do
-    rest <- Text.stripPrefix
-        "Process still running.\nsession_id: "
-        output
-    let (sessionText, withNewline) = Text.breakOn "\n" rest
-    body <- Text.stripPrefix "\n" withNewline
-    sessionId <- readWholeInt sessionText
-    pure (sessionId, body)
-
-terminalShellResult :: Text -> Maybe (BlockState, Text)
-terminalShellResult output =
-    case Text.stripPrefix "Exit code: " output of
-        Just rest ->
-            let (codeText, withNewline) = Text.breakOn "\n" rest
-            in case (readWholeInt codeText, Text.stripPrefix "\n" withNewline) of
-                (Just code, Just body) ->
-                    Just
-                        ( if code == 0 then BlockComplete else BlockFailed
-                        , body
-                        )
-                _ -> Nothing
-        Nothing
-            | "Error: Command cancelled" `Text.isPrefixOf` output ->
-                Just (toolResultState output, output)
-            | "Error: Command timed out" `Text.isPrefixOf` output ->
-                Just (toolResultState output, output)
-            | otherwise -> Nothing
-
-readWholeInt :: Text -> Maybe Int
-readWholeInt text =
-    case reads (Text.unpack text) of
-        [(value, "")] -> Just value
-        _ -> Nothing
-
-isTodoTool :: Text -> Bool
-isTodoTool name =
-    canonicalToolName name `elem` ["todo_write", "update_plan"]
-
-isShellProcessTool :: Text -> Bool
-isShellProcessTool name =
-    canonicalToolName name `elem` ["shell_command", "write_stdin"]
-
-toolBlockKind :: Text -> BlockKind
-toolBlockKind rawName
-    | name `elem` ["run_terminal_cmd", "shell_command", "write_stdin", "run_ghci", "exec"] =
-        BlockShell
-    | name `elem` ["search_replace", "apply_patch", "Write", "NotebookEdit"] =
-        BlockEdit
-    | name `elem` ["todo_write", "update_plan"] =
-        BlockTodo
-    | isInspectionTool rawName = BlockInspect
-    | otherwise = BlockTool
-  where
-    name = canonicalToolName rawName
-
--- | Syntax grammar for code carried in a shell-style tool block.
--- The title is retained alongside the source after the original call leaves
--- the live-tool map, so it also identifies exec's JavaScript code here.
-blockCodeLanguage :: UiBlock -> Maybe Text
-blockCodeLanguage block
-    | block.blockKind /= BlockShell = Nothing
-    | Text.null (Text.strip block.blockDetail) = Nothing
-    | block.blockTitle == "$ exec" = Just "javascript"
-    | otherwise = Just "haskell"
-
-outputLooksFailed :: Text -> Bool
-outputLooksFailed output =
-    let lowered = Text.toLower (Text.strip output)
-    in "error:" `Text.isPrefixOf` lowered
-        || "exit: 1" `Text.isPrefixOf` lowered
-        || "exit: 2" `Text.isPrefixOf` lowered
-
-toolResultState :: Text -> BlockState
-toolResultState output
-    | "tool call rejected by user" `Text.isInfixOf` lowered =
-        BlockDenied
-    | structuredCancellation lowered =
-        BlockCancelled
-    | outputLooksFailed output = BlockFailed
-    | Just code <- exitCodeFrom lowered
-    , code /= 0 = BlockFailed
-    | otherwise = BlockComplete
-  where
-    lowered = Text.toLower (Text.strip output)
-
-structuredCancellation :: Text -> Bool
-structuredCancellation output =
-    let header = Text.strip (headLine output)
-    in header == "exit: cancelled"
-        || header == "error: command cancelled"
-        || header == "cancelled"
-        || "cancelled (" `Text.isPrefixOf` header
-
-headLine :: Text -> Text
-headLine = Text.takeWhile (/= '\n')
-
-exitCodeFrom :: Text -> Maybe Int
-exitCodeFrom text = do
-    rest <-
-        Text.stripPrefix "exit:" text
-            <|> Text.stripPrefix "exit code:" text
-    case reads (Text.unpack (Text.takeWhile (not . isSpace) (Text.strip rest))) of
-        [(code, "")] -> Just code
-        _ -> Nothing
